+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 * (deprecation) The `installation_key` config is deprecated.
 * (improvement) Bump required versions to PHP 8.2+ and Symfony 6.3+ 
+* (improvement) Rename `live` hosting tier to `production`. 
+* (deprecation) Deprecated `live` hosting tier. 
 
 
 2.0.6